requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
-e file:///Users/bo/Work/webdev/aiohttp_json_api
aiohttp==2.0.7
async-timeout==1.2.1      # via aiohttp
attrs==17.2.0
boltons==17.1.0
chardet==3.0.3            # via aiohttp
inflection==0.3.1
jsonpointer==1.10
multidict==2.1.5          # via aiohttp, yarl
python-dateutil==2.6.0
six==1.10.0               # via python-dateutil
yarl==0.10.2              # via aiohttp
